Fix build warning on const long long to int implicit conversion.
* inspector/InspectorApplicationCacheAgent.cpp:
(WebCore::InspectorApplicationCacheAgent::buildObjectForApplicationCacheResource):

Canonical link: https://commits.webkit.org/100797@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@113472 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/inspector/InspectorApplicationCacheAgent.cpp
@@ -203,7 +203,7 @@ PassRefPtr<TypeBuilder::ApplicationCache::ApplicationCacheResource> InspectorApp
 
     RefPtr<TypeBuilder::ApplicationCache::ApplicationCacheResource> value = TypeBuilder::ApplicationCache::ApplicationCacheResource::create()
         .setUrl(resourceInfo.m_resource.string())
-        .setSize(resourceInfo.m_size)
+        .setSize(static_cast<int>(resourceInfo.m_size))
         .setType(types);
     return value;
 }
